// src/app/shared/models/characters.model.ts
export interface ICharacters {
    name: string;
    gender: string;
    origin: string;
    createdAt: string
    image: string;
}
export interface IChrDetail{
    img: string;
    currentStatus: number;
}
export interface IChrMoreDetail{
    createdAt:String;
    name:String;
    origin:String;
    gender:String;
}